Add rust-messagepack to example
This could be interesting to some audience since it shows that at least
in this case, rust-messagepack seems to be the smallest representation
out of the box.

For comparison, I gzip compressed the bincode variant and reduced size
to 349 bytes, down from 419 bytes, which makes the 206 bytes of
messagepack even more attractive. It's nearly one-third of the the
size of JSON, coming in at 587 bytes.